Add files via upload
I'm studying a book called "Get Coding Kids", which is really well taught and a lot of the amazon reviews were from adults who found it useful for adults too.  Here's an exercise I completed unassisted to design a website based on the theme of the book.  I used some a few techniques I've learned from other books such as absolute positioning to make the headers plush to the top and left of the screen so no white gaps and then creating a duplicate hidden div without absolute positioning so any further divs wouldnt be hidden under the absolute.